MajorAyushh/Backend/requirements.txt
arxiv
ddgs
duckduckgo-search
google-genai
groq
langchain-community
pandas
PyMuPDF
python-dotenv
requests
streamlit
fastapi
uvicorn[standard]
